remove unused code functions
